// prisma/generated/zod/inputTypeSchemas/BookingActivityLogUpdateInputSchema.ts
import type { Prisma } from '@prisma/client';

import { z } from 'zod';
import { StringFieldUpdateOperationsInputSchema } from './StringFieldUpdateOperationsInputSchema';
import { NullableStringFieldUpdateOperationsInputSchema } from './NullableStringFieldUpdateOperationsInputSchema';
import { BoolFieldUpdateOperationsInputSchema } from './BoolFieldUpdateOperationsInputSchema';
import { BookingActionTypeSchema } from './BookingActionTypeSchema';
import { EnumBookingActionTypeFieldUpdateOperationsInputSchema } from './EnumBookingActionTypeFieldUpdateOperationsInputSchema';
import { NullableDateTimeFieldUpdateOperationsInputSchema } from './NullableDateTimeFieldUpdateOperationsInputSchema';
import { DateTimeFieldUpdateOperationsInputSchema } from './DateTimeFieldUpdateOperationsInputSchema';
import { MerchantUpdateOneRequiredWithoutBookingActivityLogNestedInputSchema } from './MerchantUpdateOneRequiredWithoutBookingActivityLogNestedInputSchema';
import { BookingUpdateOneRequiredWithoutBookingActivityLogNestedInputSchema } from './BookingUpdateOneRequiredWithoutBookingActivityLogNestedInputSchema';

export const BookingActivityLogUpdateInputSchema: z.ZodType<Prisma.BookingActivityLogUpdateInput> = z.object({
  id: z.union([ z.string().cuid(),z.lazy(() => StringFieldUpdateOperationsInputSchema) ]).optional(),
  userId: z.union([ z.string(),z.lazy(() => NullableStringFieldUpdateOperationsInputSchema) ]).optional().nullable(),
  systemAction: z.union([ z.boolean(),z.lazy(() => BoolFieldUpdateOperationsInputSchema) ]).optional(),
  action: z.union([ z.lazy(() => BookingActionTypeSchema),z.lazy(() => EnumBookingActionTypeFieldUpdateOperationsInputSchema) ]).optional(),
  previousValue: z.union([ z.string(),z.lazy(() => NullableStringFieldUpdateOperationsInputSchema) ]).optional().nullable(),
  newValue: z.union([ z.string(),z.lazy(() => NullableStringFieldUpdateOperationsInputSchema) ]).optional().nullable(),
  description: z.union([ z.string(),z.lazy(() => NullableStringFieldUpdateOperationsInputSchema) ]).optional().nullable(),
  ipAddress: z.union([ z.string(),z.lazy(() => NullableStringFieldUpdateOperationsInputSchema) ]).optional().nullable(),
  userAgent: z.union([ z.string(),z.lazy(() => NullableStringFieldUpdateOperationsInputSchema) ]).optional().nullable(),
  performedAt: z.union([ z.coerce.date(),z.lazy(() => NullableDateTimeFieldUpdateOperationsInputSchema) ]).optional().nullable(),
  performedBy: z.union([ z.string(),z.lazy(() => NullableStringFieldUpdateOperationsInputSchema) ]).optional().nullable(),
  createdAt: z.union([ z.coerce.date(),z.lazy(() => DateTimeFieldUpdateOperationsInputSchema) ]).optional(),
  merchant: z.lazy(() => MerchantUpdateOneRequiredWithoutBookingActivityLogNestedInputSchema).optional(),
  booking: z.lazy(() => BookingUpdateOneRequiredWithoutBookingActivityLogNestedInputSchema).optional()
}).strict();

export default BookingActivityLogUpdateInputSchema;